refactor: Days 컴포넌트 분리
